Add feature to toggle pause songs
Note: Given how the counting of seconds and thus the progress bar works,
the song and the progress bar can actually desync because of the
pausing. This is really only noticeable if the user spams pause and
unpause, though it's something to keep in mind.

Given that this will likely not be an issue, it won't be fixed for the
time being.